SearchAdapter + SearchResults Combined
SearchAdapter and SearchResults have been combined into one file
